git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@4436 19eef36c-86f0-de11-afab-0004ac968479

+++ code/src/main/java/com/algoTrader/util/LongMap.java
@@ -12,7 +12,7 @@ public class LongMap<K> extends ConcurrentHashMap<K, AtomicLong> {
 
     private static final long serialVersionUID = -847488464256946086L;
 
-    public void increment(K key, int value) {
+    public void increment(K key, long value) {
 
         super.putIfAbsent(key, new AtomicLong(0));
         super.get(key).addAndGet(value);
